+++ client/src/components/profile.ts
@@ -88,6 +88,7 @@ export interface SSHProfile extends BaseProfile {
   saspath: string;
   port: number;
   username: string;
+  privateKeyFilePath?: string;
 }
 
 export interface COMProfile extends BaseProfile {
@@ -580,6 +581,15 @@ export class ProfileConfig {
         return;
       }
 
+      const keyPath = await createInputTextBox(
+        ProfilePromptType.PrivateKeyFilePath,
+        profileClone.privateKeyFilePath,
+      );
+
+      if (keyPath) {
+        profileClone.privateKeyFilePath = keyPath;
+      }
+
       await this.upsertProfile(name, profileClone);
     } else if (profileClone.connectionType === ConnectionType.COM) {
       profileClone.sasOptions = [];
@@ -657,6 +667,7 @@ export enum ProfilePromptType {
   SASPath,
   Port,
   Username,
+  PrivateKeyFilePath,
 }
 
 /**
@@ -793,6 +804,11 @@ const input: ProfilePromptInput = {
     title: l10n.t("SAS Server Username"),
     placeholder: l10n.t("Enter your username"),
     description: l10n.t("Enter your SAS server username."),
+  },
+  [ProfilePromptType.PrivateKeyFilePath]: {
+    title: l10n.t("Private Key File Path (optional)"),
+    placeholder: l10n.t("Enter the local private key file path"),
+    description: l10n.t("To use the SSH Agent or a password, leave blank."),
   },
 };