+++ client/src/connection/itc/index.ts
@@ -141,6 +141,7 @@ export class ITCSession extends Session {
         `$runner.Setup($profileHost,$username,$password,$port,$protocol,$serverName,$displayLang)\n`,
         this.onWriteComplete,
       );
+      this._workDirectoryParser.reset();
       this._shellProcess.stdin.write(
         "$runner.ResolveSystemVars()\n",
         this.onWriteComplete,
@@ -514,6 +515,7 @@ export class ITCSession extends Session {
     const globalStorageUri = getGlobalStorageUri();
     try {
       await workspace.fs.readDirectory(globalStorageUri);
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
     } catch (e) {
       await workspace.fs.createDirectory(globalStorageUri);
     }