+++ client/src/connection/rest/util.ts
@@ -55,7 +55,8 @@ export const resourceType = (item: ContentItem): string | undefined => {
   if (!isValidItem(item)) {
     return;
   }
-  const { write, delete: del, addMember } = getPermission(item);
+
+  const { write, delete: del, addMember } = item.permission;
   const isRecycled = isItemInRecycleBin(item);
   const actions = [
     addMember && !isRecycled && "createChild",
@@ -94,9 +95,16 @@ export const resourceType = (item: ContentItem): string | undefined => {
   return actions.sort().join("-");
 };
 
-export const getUri = (item: ContentItem, readOnly?: boolean): Uri =>
+export const getSasContentUri = (item: ContentItem, readOnly?: boolean): Uri =>
   Uri.parse(
     `${readOnly ? `${ContentSourceType.SASContent}ReadOnly` : ContentSourceType.SASContent}:/${
+      item.name
+    }?id=${getResourceIdFromItem(item)}`,
+  );
+
+export const getSasServerUri = (item: ContentItem, readOnly?: boolean): Uri =>
+  Uri.parse(
+    `${readOnly ? `${ContentSourceType.SASServer}ReadOnly` : ContentSourceType.SASServer}:/${
       item.name
     }?id=${getResourceIdFromItem(item)}`,
   );