+++ server/src/sas/LexerEx.ts
@@ -3055,7 +3055,6 @@ export class LexerEx {
               });
               break;
             } // not break
-          // eslint-disable-next-line no-fallthrough
           default: {
             this.tryToHandleSectionEnd_(token);
             let generalProcStmt = true;
@@ -3297,7 +3296,6 @@ export class LexerEx {
               });
               break;
             } // attention: not break here
-          // eslint-disable-next-line no-fallthrough
           default:
             if (Lexer.isCards[word]) {
               this.cardsState = this.CARDS_STATE.IN_CMD;